Moved useDatesListFilterState; js_only.
