+++ assets/prototype/domain/eventEditor/datetimes/dateList/filter/DatesListFilterBar.js
@@ -5,25 +5,14 @@ import DatesListEntityFilters from './DatesListEntityFilters';
  * DatesListFilterBar
  * filters for controlling the display of a list of Event Dates
  *
- * @param {string} showDates
- * @param {string} datesSortedBy
- * @param {string} displayDates
- * @param {string} searchText
- * @param {number} perPage
- * @param {string} view
 
  * @return {Object} EditorDatesListView with added DateListFilters
  */
-const DatesListFilterBar = ({ datetimes }) => {
+const DatesListFilterBar = () => {
 	const listId = 'event-editor-dates-list';
+	const entityFilters = <DatesListEntityFilters listId={listId} />;
 
-	return (
-		<EntityListFilterBar
-			datetimes={datetimes}
-			listId={listId}
-			entityFilters={<DatesListEntityFilters listId={listId} />}
-		/>
-	);
+	return <EntityListFilterBar listId={listId} entityFilters={entityFilters} />;
 };
 
 export default DatesListFilterBar;
